EV0053
EV0053 is one of 4 unused cutscenes that were used as placeholders for the FMV of the Egg Carrier crashing. This one is for Tails' story.

The original file had Japanese text (Which I've left in a comment), but it wouldn't show up in-game and I didn't feel like screwing around with encoding for a small thing like this.