Major Variable Overhaul
All object variables are global now (like the game does) so FreeTask actually cleans data now. Previously they were all local and set to 0 at the beginning of the event.

Everything should be reset at the end of the event, so there shouldn't be problems with tasks not being freed, but let me know if there's any issues.

Eventually I'll probably collapse the numerous clones of Amy's bird etc. into one thing, but that'd be annoying to do before the rest of the events are done.